Move charge reference data into charge version A charge version always has to have a charge reference so it makes sense to include it in the charge version data
